+++ core/src/nucleus/ribosome/api/sign.rs
@@ -1,5 +1,5 @@
 use crate::nucleus::ribosome::{api::ZomeApiResult, Runtime};
-use holochain_core_types::{error::HcResult, signature::Signature};
+use holochain_core_types::{error::HcResult, json::JsonString, signature::Signature};
 use holochain_dpki::keypair::generate_random_sign_keypair;
 use holochain_sodium::secbuf::SecBuf;
 use holochain_wasm_utils::api_serialization::sign::{OneTimeSignArgs, SignArgs, SignOneTimeResult};
@@ -28,7 +28,9 @@ pub fn invoke_sign(runtime: &mut Runtime, args: &RuntimeArgs) -> ZomeApiResult {
         }
     };
 
-    let signature = context.sign(sign_args.payload.clone());
+    let signature = context
+        .sign(sign_args.payload.clone())
+        .map(|sig| JsonString::from_json(&sig));
 
     context.log(format!(
         "debug/zome: signature of data:{:?} by:{:?} is:{:?}",
@@ -99,7 +101,7 @@ mod test_super {
             ZomeApiFunction::Sign.as_str(),
             r#"{ "payload": "this is data" }"#.as_bytes().to_vec(),
         );
-        assert_eq!(JsonString::from(r#"{"ok":true,"value":"xoEEoLF1yWM4VBNtjEwrfM/iVzjuAxxbkOyBWi0LV0+1CAH/PCs9MErnbmFeZRtQNtw7+SmVrm7Irac4lZsaDA==","error":"null"}"#), call_result,);
+        assert_eq!(JsonString::from_json(r#"{"ok":true,"value":"xoEEoLF1yWM4VBNtjEwrfM/iVzjuAxxbkOyBWi0LV0+1CAH/PCs9MErnbmFeZRtQNtw7+SmVrm7Irac4lZsaDA==","error":"null"}"#), call_result,);
     }
 
     #[test]